Some tweaks
