+++ src/app/Applications/QuickTime/QuickTimeContext.tsx
@@ -1,7 +1,7 @@
 import { QuickTimeAppInfo } from '@/app/Applications/QuickTime/QuickTimeMoviePlayer'
 import { classicyAppEventHandler, ClassicyStore } from '@/app/SystemFolder/ControlPanels/AppManager/ClassicyAppManager'
 
-type classicyQuickTimeDocument = {
+export type ClassicyQuickTimeDocument = {
     url: string
     name?: string
     options?: any
@@ -10,8 +10,8 @@ type classicyQuickTimeDocument = {
 }
 type classicyQuickTimeEvent = {
     type: string
-    document?: classicyQuickTimeDocument
-    documents?: classicyQuickTimeDocument[]
+    document?: ClassicyQuickTimeDocument
+    documents?: ClassicyQuickTimeDocument[]
 }
 
 export const classicyQuickTimeEventHandler = (ds: ClassicyStore, action: classicyQuickTimeEvent) => {
@@ -22,12 +22,15 @@ export const classicyQuickTimeEventHandler = (ds: ClassicyStore, action: classic
             app: QuickTimeAppInfo,
         })
     }
+
     if (!ds.System.Manager.App.apps[appId]?.hasOwnProperty('data')) {
         ds.System.Manager.App.apps[appId].data = {}
     }
+
     if (!ds.System.Manager.App.apps[appId]?.data?.hasOwnProperty('openFiles')) {
         ds.System.Manager.App.apps[appId].data['openFiles'] = []
     }
+
     const openDocUrls = ds.System.Manager.App.apps[appId].data['openFiles'].map((app) => app.url)
 
     switch (action.type) {
@@ -36,20 +39,31 @@ export const classicyQuickTimeEventHandler = (ds: ClassicyStore, action: classic
                 ds.System.Manager.App.apps[appId].data['openFiles'] = Array.from(
                     new Set([...ds.System.Manager.App.apps[appId].data['openFiles'], action.document])
                 )
+                ds = classicyAppEventHandler(ds, {
+                    type: 'ClassicyAppOpen',
+                    app: QuickTimeAppInfo,
+                })
             }
             break
         }
         case 'ClassicyAppQuickTimeOpenDocuments': {
-            const docs = action.documents.filter((doc) => !openDocUrls.includes(doc.url))
+            const docs = action.documents?.filter((doc) => !openDocUrls.includes(doc.url))
+            if (!docs) {
+                break
+            }
             ds.System.Manager.App.apps[appId].data['openFiles'] = Array.from(
                 new Set([...ds.System.Manager.App.apps[appId].data['openFiles'], ...docs])
             )
+            ds = classicyAppEventHandler(ds, {
+                type: 'ClassicyAppOpen',
+                app: QuickTimeAppInfo,
+            })
             break
         }
         case 'ClassicyAppQuickTimeCloseDocument': {
             ds.System.Manager.App.apps[appId].data['openFiles'] = ds.System.Manager.App.apps[appId].data[
                 'openFiles'
-            ].filter((p: classicyQuickTimeDocument) => p.url != action.document.url)
+            ].filter((p: ClassicyQuickTimeDocument) => p.url != action.document.url)
             break
         }
     }
