+++ src/app/SystemFolder/SystemResources/App/ClassicyAppUtils.tsx
@@ -8,3 +8,18 @@ export const quitAppHelper = (appId: string, appName: string, appIcon: string) =
         },
     }
 }
+
+export const quitMenuItemHelper = (appId: string, appName: string, appIcon: string) => {
+    return {
+        id: appId + '_quit',
+        title: 'Quit',
+        event: 'ClassicyAppClose',
+        eventData: {
+            app: {
+                id: appId,
+                title: appName,
+                icon: appIcon,
+            },
+        },
+    }
+}
